requirements.txt
azure-identity
azure-ai-projects
python-dotenv
azure-functions
requests